+++ src/integral.h
@@ -1,3 +1,4 @@
+#include <stddef.h>
 #include <stdbool.h>
 
 typedef signed QS_INTEGRAL_POWERTYPE QsPower;
@@ -5,9 +6,9 @@ typedef unsigned QsPrototype;
 typedef struct QsIntegral* QsIntegral;
 
 QsIntegral qs_integral_new_from_string( const char* );
-QsIntegral qs_integral_new_from_binary( const char*,unsigned );
-unsigned qs_integral_print( const QsIntegral,char** );
-unsigned qs_integral_to_binary( QsIntegral i,char** out );
+QsIntegral qs_integral_new_from_binary( const char*,size_t );
+size_t qs_integral_print( const QsIntegral,char** );
+size_t qs_integral_to_binary( QsIntegral i,char** out );
 bool qs_integral_cmp( const QsIntegral,const QsIntegral );
 QsIntegral qs_integral_cpy( const QsIntegral );
 const QsPower* qs_integral_powers( const QsIntegral );
